refactor(electron): rename bridge/core.ts to bridge/BridgeFactory.ts
Rename bridge/core.ts to bridge/BridgeFactory.ts for better semantic clarity.
Update import path in bridge/index.ts accordingly.

This aligns with the file's purpose as a factory for creating core bridge instances.
All tests continue to pass (469/469) and validation suite clean.
